requirements.txt
django==3.2.16
django-cors-headers
django-cities-light
python-dateutil==2.8.2
Pillow
mysqlclient==2.2.0
reportlab==3.6.5
xlwt
django-admin-interface
django-notifications-hq
django-jsignature
django_apscheduler
requests==2.26.0
django-import-export
pandas
python-dotenv
django-crontab
matplotlib
django-user-activity-log2==0.0.25